test: shared makeTestEnv/makeLeagueTestHarness/withCookie (feedback #9)
The 9-copy testEnv literal, the 3-copy fixed-clock app trio + boundary
constants, and withCookie now live in test/setup/{test-env,league-app}.ts;
test bodies and assertions untouched.
