+++ apps/api/test/discovery.test.ts
@@ -1,48 +1,13 @@
 import { afterAll, beforeEach, describe, expect, it } from "vitest";
-import { createDb } from "@picksleagues/db";
-import { FixedClock, type Env } from "@picksleagues/core";
 import { LEAGUE_STATUS, LEAGUE_VISIBILITY, type DiscoveryResponse } from "@picksleagues/schemas";
-import { createApp } from "../src/app";
-import { createAuth } from "../src/auth";
 import { createAuthenticatedUser } from "./setup/auth-helpers";
 import { insertLeague, seedSeason } from "./setup/league-helpers";
+import { makeLeagueTestHarness, WEEK1_KICKOFF, withCookie } from "./setup/league-app";
 import { resetDb } from "./setup/reset-db";
-import { getTestDatabaseUrl } from "./setup/test-database-url";
 
-const testEnv: Env = {
-  APP_ENV: "local",
-  DATABASE_URL: getTestDatabaseUrl(),
-  BETTER_AUTH_SECRET: "a".repeat(32),
-  BETTER_AUTH_URL: "http://localhost:3000",
-  GOOGLE_CLIENT_ID: "google-id",
-  GOOGLE_CLIENT_SECRET: "google-secret",
-  DISCORD_CLIENT_ID: "discord-id",
-  DISCORD_CLIENT_SECRET: "discord-secret",
-  JOB_SECRET: "b".repeat(32),
-  ADMIN_USER_IDS: [],
-};
-
-const WEEK1_KICKOFF = new Date("2026-09-13T17:00:00.000Z");
-// Two apps over the same DB: one with the clock before the seeded kickoff,
-// one after — discovery excludes a league once its start has passed, and the
-// cutoff is derived per-request from the clock (arch §Locking Model).
-const PRE_START_NOW = new Date("2026-09-01T00:00:00.000Z");
-const POST_START_NOW = new Date("2026-09-13T17:00:00.001Z");
-
-const db = createDb(getTestDatabaseUrl());
-const auth = createAuth({ env: testEnv, db });
-const app = createApp({ auth, db, clock: async () => new FixedClock(PRE_START_NOW) });
-const appAfterKickoff = createApp({
-  auth,
-  db,
-  clock: async () => new FixedClock(POST_START_NOW),
-});
+const { db, auth, app, appAfterKickoff } = makeLeagueTestHarness();
 
 type App = typeof app;
-
-function withCookie(cookie: string | undefined): Record<string, string> {
-  return cookie ? { cookie } : {};
-}
 
 function getDiscovery(cookie: string | undefined, q?: string, on: App = app) {
   const query = q !== undefined ? `?q=${q}` : "";
